# server/api/swagger.yaml
openapi: 3.0.1
info:
  title: MMS Example API
  description: Documentation for MMS API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.txt
  version: 0.0.1
servers:
- url: https://mms.example.com
  description: Generated server url
security:
- bearerToken: []
paths:
...
  /projects/{projectId}/refs/{refId}/views/{viewId}:
    get:
      tags:
      - Views
      operationId: getView
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: viewId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
  /projects/{projectId}/refs/{refId}/mounts:
    get:
      tags:
      - Views
      operationId: getMounts
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MountsResponse'
  /projects/{projectId}/refs/{refId}/views:
    put:
      tags:
      - Views
      operationId: getViews
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
    post:
      tags:
      - Views
      operationId: createOrUpdateViews
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: overwrite
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
  /projects/{projectId}/refs/{refId}/documents:
    get:
      tags:
      - Views
      operationId: getDocuments
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentsResponse'
  /projects/{projectId}/refs/{refId}/groups:
    get:
      tags:
      - Views
      operationId: getGroups
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupsResponse'
...
  /projects/{projectId}/refs/{refId}/elements/{elementId}:
    get:
      tags:
      - Elements
      operationId: getArtifactForElement_1
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: elementId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: Accept
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
    post:
      tags:
      - Artifacts
      operationId: createOrUpdateArtifacts
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: elementId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/elements_elementId_body'
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
    delete:
      tags:
      - Elements
      operationId: deleteElement
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: elementId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsCommitResponse'
...
  /projects/{projectId}/refs/{refId}/commits:
    get:
      tags:
      - Commits
      operationId: getRefCommits
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: maxTimestamp
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitsResponse'
  /projects/{projectId}/commits/{commitId}:
    get:
      tags:
      - Commits
      operationId: getCommit
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: commitId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitsResponse'
  /projects/{projectId}/refs/{refId}/elements/{elementId}/commits:
    get:
      tags:
      - Commits
      operationId: getElementCommits
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: elementId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitsResponse'
  /projects/{projectId}/commits:
    put:
      tags:
      - Commits
      operationId: getCommits
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitsRequest'
        required: true
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitsResponse'
  /projects/{projectId}/refs/{refId}/elements:
    get:
      tags:
      - Elements
      operationId: getAllElements
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
            application/x-ndjson: {}
    put:
      tags:
      - Elements
      operationId: getElements
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: commitId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
    post:
      tags:
      - Elements
      operationId: createOrUpdateElements
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: overwrite
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElementsRequest'
        required: true
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsCommitResponse'
    delete:
      tags:
      - Elements
      operationId: deleteElements
      parameters:
      - name: projectId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: refId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
...
components:
  schemas:
...
    Element:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        type:
          type: string
        _inRefIds:
          type: array
          readOnly: true
          items:
            type: string
            readOnly: true
        _created:
          type: string
          readOnly: true
        _projectId:
          type: string
          readOnly: true
        _refId:
          type: string
          readOnly: true
        _commitId:
          type: string
          readOnly: true
        _creator:
          type: string
          readOnly: true
        _docId:
          type: string
          readOnly: true
        _modified:
          type: string
          readOnly: true
        _modifier:
          type: string
          readOnly: true
      additionalProperties:
        type: object
      x-schema-name: Element
    ElementsResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        elements:
          type: array
          items:
            $ref: '#/components/schemas/Element'
      example:
        rejected:
        - code: 0
          message: message
          object: {}
        - code: 0
          message: message
          object: {}
        elements:
        - key: {}
        - key: {}
        messages:
        - messages
        - messages
    Rejection:
      type: object
      properties:
        object:
          type: object
        code:
          type: integer
          format: int32
        message:
          type: string
      nullable: true
      example:
        code: 0
        message: message
        object: {}
    Mount:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        type:
          type: string
        _created:
          type: string
          readOnly: true
        _projectId:
          type: string
          readOnly: true
        _refId:
          type: string
          readOnly: true
        _commitId:
          type: string
          readOnly: true
        _creator:
          type: string
          readOnly: true
        _docId:
          type: string
          readOnly: true
        _modified:
          type: string
          readOnly: true
        _modifier:
          type: string
          readOnly: true
      additionalProperties:
        type: object
      x-schema-name: Mount
    MountsResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Mount'
      example:
        projects:
        - key: {}
        - key: {}
        rejected:
        - code: 0
          message: message
          object: {}
        - code: 0
          message: message
          object: {}
        messages:
        - messages
        - messages
    ElementsRequest:
      required:
      - elements
      type: object
      properties:
        source:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        elements:
          type: array
          items:
            $ref: '#/components/schemas/Element'
    DocumentsResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Element'
      example:
        rejected:
        - code: 0
          message: message
          object: {}
        - code: 0
          message: message
          object: {}
        documents:
        - key: {}
        - key: {}
        messages:
        - messages
        - messages
    GroupsResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Element'
      example:
        rejected:
        - code: 0
          message: message
          object: {}
        - code: 0
          message: message
          object: {}
        messages:
        - messages
        - messages
        groups:
        - key: {}
        - key: {}
...
    Commit:
      required:
      - id
      type: object
      properties:
        comment:
          type: string
          readOnly: true
        deleted:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/ElementVersion'
        added:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/ElementVersion'
        updated:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/ElementVersion'
        source:
          type: string
          readOnly: true
        id:
          type: string
        _created:
          type: string
          readOnly: true
        _projectId:
          type: string
          readOnly: true
        _refId:
          type: string
          readOnly: true
        _commitId:
          type: string
          readOnly: true
        _creator:
          type: string
          readOnly: true
        _docId:
          type: string
          readOnly: true
      additionalProperties:
        required:
        - id
        type: object
      x-schema-name: Commit
    CommitsResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        commits:
          type: array
          items:
            $ref: '#/components/schemas/Commit'
      example:
        rejected:
        - code: 0
          message: message
          object: {}
        - code: 0
          message: message
          object: {}
        messages:
        - messages
        - messages
        commits:
        - key: {}
        - key: {}
    ElementVersion:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        _previousDocId:
          type: string
          readOnly: true
        _docId:
          type: string
          readOnly: true
      readOnly: true
    CommitsRequest:
      required:
      - commits
      type: object
      properties:
        commits:
          type: array
          items:
            $ref: '#/components/schemas/Commit'
    ElementsCommitResponse:
      type: object
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
            nullable: true
        rejected:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Rejection'
        elements:
          type: array
          items:
            $ref: '#/components/schemas/Element'
        commitId:
          type: string
          nullable: true
      example:
        rejected:
        - code: 0
          message: message
          object: {}
        - code: 0
          message: message
          object: {}
        elements:
        - key: {}
        - key: {}
        messages:
        - messages
        - messages
        commitId: commitId
...
    inline_response_200:
      oneOf:
      - $ref: '#/components/schemas/ElementsResponse'
      - type: string
    elements_elementId_body:
      type: object
      properties:
        file:
          type: string
          format: binary
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT