# server/go/model_element_version.go
/*
 * MMS Example API
 *
 * Documentation for MMS API
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type ElementVersion struct {
	Id string `json:"id,omitempty"`

	Type_ string `json:"type,omitempty"`

	PreviousDocId string `json:"_previousDocId,omitempty"`

	DocId string `json:"_docId,omitempty"`
}