# server/go/model_elements_request.go
/*
 * MMS Example API
 *
 * Documentation for MMS API
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type ElementsRequest struct {
	Source string `json:"source,omitempty"`

	Comment string `json:"comment,omitempty"`

	Elements []map[string]interface{} `json:"elements"`
}