# server/go/model_groups_response.go
/*
 * MMS Example API
 *
 * Documentation for MMS API
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type GroupsResponse struct {
	Messages []string `json:"messages,omitempty"`

	Rejected []Rejection `json:"rejected,omitempty"`

	Groups []map[string]interface{} `json:"groups,omitempty"`
}